chore: move types into context to avoid more circ deps
