+++ packages/ibm-products/src/components/InterstitialScreen/InterstitialScreenFooter.tsx
@@ -8,8 +8,12 @@ import React, { useContext, useEffect, useMemo, useRef, useState } from 'react';
 
 import PropTypes from 'prop-types';
 
-import { ActionType, disableButtonConfigType } from './InterstitialScreen';
-import { InterstitialScreenContext, blockClass } from './context';
+import {
+  InterstitialScreenContext,
+  blockClass,
+  ActionType,
+  disableButtonConfigType,
+} from './context';
 import { Button, InlineLoading, ModalFooter } from '@carbon/react';
 import { clamp } from '../../global/js/utils/clamp';
 import { ArrowRight } from '@carbon/react/icons';
